Cleans routing
- Merges and moves Navigator-based routing to ./js/router
- Replaces default navbar with Header component
